+++ changelog.md
@@ -9,6 +9,7 @@
 - Further improved "Optimize Bullet Trails" option to save even more RAM
 - Added "Optimize Linear Cast Pool" to fix a memory leak in physics calculations
 - Optimize more various math functions
+- Added "Optimize Pixel Movement Gen"
 
 ## 1.3.3 (2025-06-01)